Make loading a little feel more snappy
It's not necessarily faster, but it feels faster because it actually updates the page (and shows a progress bar) when you click on the page.

Oh, and if the user clicks on a different page while the page is loading, it'll actually stop prematurely, so it's not still loading when it can't even put the cards on screen.